# docs/openapi.yaml
openapi: 3.0.3
info:
  title: StudyOwl Auth Service API
  version: 0.4.0
  description: |
    Thin backend layer in front of Amazon Cognito that handles phone-number
    or email-address login/registration for the StudyOwl app.

    Design notes (see docs/studyowl-auth-flow.md and auth-service/README.md
    for the full write-up):

    - There is deliberately **no** "does this identifier exist" endpoint.
      Existence is checked server-side (Cognito `AdminGetUser`) inside
      `request-otp`, which always returns the same generic response either
      way — exposing that check directly would let anyone enumerate
      registered phone numbers or email addresses.
    - New-vs-returning only becomes visible in the `confirm-otp` response,
      and only after the caller has proven they hold the OTP.
    - Built on Cognito's native choice-based `USER_AUTH` flow, with either
      the `SMS_OTP` or `EMAIL_OTP` challenge depending on whether `identifier`
      is a phone number or an email address — no Lambda triggers, no
      passwords anywhere (not even a throwaway one). `request-otp` returns a
      `session` token (opaque, generated by Cognito itself) that the client
      must pass back unmodified to `confirm-otp` — this is Cognito's own
      session mechanism, not a custom correlation ID this service invents or
      stores.
    - Because the native flow issues real tokens the moment the code checks
      out, `confirm-otp` returns tokens for new and returning users alike —
      there's no way to withhold tokens from a new user until they finish
      `complete-profile`, unlike an earlier password-based version of this
      service. `complete-profile` is purely about finishing registration
      (name/photo), not about token issuance.
    - `request-otp` should be rate-limited per identifier (e.g. 3/hour)
      since it is unauthenticated and triggers a real SMS/email send.
    - Session tokens are Cognito ID/access/refresh tokens; this service does
      not mint its own JWTs.
    - Requires the Cognito User Pool to be on the Essentials or Plus feature
      plan (not Lite), with `SMS_OTP` and `EMAIL_OTP` enabled under
      choice-based sign-in, the App Client to allow `ALLOW_USER_AUTH`, and
      (for `EMAIL_OTP` specifically) the pool's email configuration set to
      use Amazon SES rather than Cognito's default email. See README.md.
    - `complete-profile` and its `photo-upload-url` step require a Bearer
      access token (the one `confirm-otp` just issued) instead of a
      request-body token. This service does **not** verify that token itself
      — it trusts an `X-User-Sub` header instead, injected by whatever
      gateway sits in front of it (Kong locally, in `kong/kong.yml`) only
      after that gateway has verified the token's signature. This is only
      safe because auth-service is never reachable except through the
      gateway; see README.md for the full reasoning.
    - Not yet implemented, worth adding for a production-grade deployment:
      device attestation (e.g. Google Play Integrity) on `request-otp`, so the
      endpoint can only be called from the genuine app, not a script hitting
      the API directly. Rate limiting alone doesn't stop an attacker who
      rotates identifiers.

servers:
  - url : http://localhost:8080/v1
    description: dev
  - url: https://api.studyowl.app/v1
    description: Production
  - url: https://api-staging.studyowl.app/v1
    description: Staging

tags:
  - name: OTP
    description: Phone number or email entry and OTP verification
  - name: Registration
    description: First-time profile setup for new users
  - name: Session
    description: Token refresh

paths:
  /auth/request-otp:
    post:
      tags: [OTP]
      summary: Request an OTP for a phone number or email address
      description: |
        Always returns the same generic response whether the identifier is
        new or already registered. Internally the backend creates a Cognito
        user with no password if the identifier is new, then starts (or
        resumes) the USER_AUTH flow with SMS_OTP or EMAIL_OTP — whichever
        matches the identifier's format — and Cognito sends the code either way.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestOtpRequest'
      responses:
        '200':
          description: OTP sent (generic — does not reveal new vs. existing)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequestOtpResponse'
        '400':
          description: Invalid identifier
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit exceeded for this identifier
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RateLimitError'

  /auth/confirm-otp:
    post:
      tags: [OTP]
      summary: Verify the OTP for a phone number or email address
      description: |
        Issues real Cognito tokens on success, for both new and returning
        users. For a new user, the app should still collect name/photo via
        `complete-profile` before treating registration as complete, even
        though it already has valid tokens — call it with the `accessToken`
        from this response as a Bearer token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmOtpRequest'
      responses:
        '200':
          description: OTP verified, tokens issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfirmOtpResponse'
        '400':
          description: Incorrect or expired OTP
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Too many incorrect attempts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RateLimitError'

  /auth/complete-profile/photo-upload-url:
    post:
      tags: [Registration]
      summary: Get a presigned S3 URL for the optional profile photo
      description: |
        Call before `complete-profile` only if the user is attaching a
        photo. The file is uploaded directly to S3 by the client; the
        resulting `photoUrl` is then passed to `complete-profile`.
      parameters:
        - $ref: '#/components/parameters/XUserSub'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [contentType]
              properties:
                contentType:
                  type: string
                  example: image/jpeg
      responses:
        '200':
          description: Presigned upload URL issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PhotoUploadUrlResponse'
        '401':
          description: Missing X-User-Sub header
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/complete-profile:
    post:
      tags: [Registration]
      summary: Submit registration details for a new user
      description: |
        Only valid for the new-user path (isNewUser: true from confirm-otp).
        Hands the profile off to the User Profile service. Does NOT issue
        tokens — confirm-otp already did that.
      parameters:
        - $ref: '#/components/parameters/XUserSub'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompleteProfileRequest'
      responses:
        '200':
          description: Profile created
        '400':
          description: Missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Missing X-User-Sub header
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/refresh-token:
    post:
      tags: [Session]
      summary: Exchange a refresh token for a new access/ID token
      description: |
        Thin pass-through to Cognito's REFRESH_TOKEN_AUTH flow. If the app
        calls Cognito directly via the AWS SDK for this step instead, this
        endpoint is not needed — confirm which approach StudyOwl is using
        before implementing it.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: New tokens issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Refresh token expired or revoked — caller must fall back to identifier-entry flow
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  parameters:
    XUserSub:
      name: X-User-Sub
      in: header
      required: true
      description: >-
        The caller's verified Cognito `sub`. Not something the client sets —
        the client sends its confirm-otp accessToken as a normal `Authorization:
        Bearer` header to whatever gateway is in front of this service (Kong
        locally; see kong/kong.yml at the repo root), and that gateway verifies
        it and injects this header before the request ever reaches auth-service.
        This service trusts the header as-is and never parses a JWT itself —
        see README.md for why that's safe only because auth-service is never
        reachable except through the gateway.
      schema:
        type: string

  schemas:
    RequestOtpRequest:
      type: object
      required: [identifier]
      properties:
        identifier:
          type: string
          description: >-
            An E.164-format phone number or an email address. The backend
            detects which one this is and requests the matching OTP challenge
            (SMS_OTP or EMAIL_OTP) accordingly.
          pattern: '^\+[1-9]\d{1,14}$|^[^\s@]+@[^\s@]+$'
          example: '+919876543210'

    RequestOtpResponse:
      type: object
      properties:
        message:
          type: string
          example: OTP sent
        otpExpirySeconds:
          type: integer
          example: 300
        session:
          type: string
          description: >-
            Cognito's own session token for the challenge in progress.
            Opaque; pass back unmodified in confirm-otp.

    ConfirmOtpRequest:
      type: object
      required: [identifier, otp, session]
      properties:
        identifier:
          type: string
          description: Must be the same value passed to request-otp.
          pattern: '^\+[1-9]\d{1,14}$|^[^\s@]+@[^\s@]+$'
        otp:
          type: string
          description: >-
            Digit-only code Cognito sent via SMS or email. Length isn't fixed
            by Cognito's own contract, so this accepts a range rather than
            one specific length.
          pattern: '^\d{4,10}$'
        session:
          type: string
          description: The session value returned by request-otp

    ConfirmOtpResponse:
      type: object
      required: [isNewUser, tokens]
      properties:
        isNewUser:
          type: boolean
        tokens:
          $ref: '#/components/schemas/Tokens'

    CompleteProfileRequest:
      type: object
      required: [firstName, lastName, role]
      description: >-
        userId/phoneNumber/email aren't collected here — userId comes from the
        caller's verified identity (X-User-Sub) and phoneNumber/email from
        Cognito, not user input. role is just the initial value (defaults to
        STUDENT server-side if omitted) — it's a mutable user-profile-service
        field from here on, not fixed for the life of the account; see that
        service's PATCH /profiles/me for how it's changed later. Otherwise
        mirrors the fields user-profile-service's own ProfileRequest accepts
        (see that service's docs/openapi.yaml).
      properties:
        firstName:
          type: string
        lastName:
          type: string
        role:
          $ref: '#/components/schemas/Role'
        photoUrl:
          type: string
          description: S3 URL returned after uploading via the presigned URL from /auth/complete-profile/photo-upload-url

    Role:
      type: string
      enum: [STUDENT, OWNER]

    PhotoUploadUrlResponse:
      type: object
      properties:
        uploadUrl:
          type: string
          description: Presigned S3 PUT URL
        photoUrl:
          type: string
          description: Final public/stored URL to send back in complete-profile once the upload succeeds
        expiresInSeconds:
          type: integer
          example: 300

    Tokens:
      type: object
      properties:
        idToken:
          type: string
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
          description: Access token lifetime in seconds

    TokenResponse:
      allOf:
        - $ref: '#/components/schemas/Tokens'

    Error:
      type: object
      properties:
        code:
          type: string
          example: INVALID_OTP
        message:
          type: string

    RateLimitError:
      allOf:
        - $ref: '#/components/schemas/Error'
        - type: object
          properties:
            retryAfterSeconds:
              type: integer
              example: 3600